Upgrade to Rust 2024 edition (#774)
# Objective

We're still on good ol' Rust 2021 edition. It's time to move on to the 2024 edition!

## Solution

Upgrade the edition and fix any errors and lints that rustc and clippy complain about. The biggest ones are if-let chains and some `use<...>` bounds.

+++ src/data_structures/sparse_secondary_map.rs
@@ -164,10 +164,10 @@ impl<V, S: hash::BuildHasher> SparseSecondaryEntityMap<V, S> {
     /// the entity was not previously removed.
     #[inline]
     pub fn remove(&mut self, entity: Entity) -> Option<V> {
-        if let hash_map::Entry::Occupied(entry) = self.slots.entry(entity.index()) {
-            if entry.get().generation == entity.generation() {
-                return Some(entry.remove_entry().1.value);
-            }
+        if let hash_map::Entry::Occupied(entry) = self.slots.entry(entity.index())
+            && entry.get().generation == entity.generation()
+        {
+            return Some(entry.remove_entry().1.value);
         }
 
         None
@@ -198,7 +198,7 @@ impl<V, S: hash::BuildHasher> SparseSecondaryEntityMap<V, S> {
     #[inline]
     pub unsafe fn get_unchecked(&self, entity: Entity) -> &V {
         debug_assert!(self.contains(entity));
-        self.get(entity).unwrap_unchecked()
+        unsafe { self.get(entity).unwrap_unchecked() }
     }
 
     /// Returns a mutable reference to the value corresponding to the entity.
@@ -220,7 +220,7 @@ impl<V, S: hash::BuildHasher> SparseSecondaryEntityMap<V, S> {
     #[inline]
     pub unsafe fn get_unchecked_mut(&mut self, entity: Entity) -> &mut V {
         debug_assert!(self.contains(entity));
-        self.get_mut(entity).unwrap_unchecked()
+        unsafe { self.get_mut(entity).unwrap_unchecked() }
     }
 
     /// Returns the value corresponding to the entity if it exists, otherwise inserts
@@ -307,11 +307,13 @@ impl<V, S: hash::BuildHasher> SparseSecondaryEntityMap<V, S> {
         entities: [Entity; N],
     ) -> [&mut V; N] {
         // Safe, see get_disjoint_mut.
-        let mut ptrs: [MaybeUninit<*mut V>; N] = MaybeUninit::uninit().assume_init();
-        for i in 0..N {
-            ptrs[i] = MaybeUninit::new(self.get_unchecked_mut(entities[i]));
+        unsafe {
+            let mut ptrs: [MaybeUninit<*mut V>; N] = MaybeUninit::uninit().assume_init();
+            for i in 0..N {
+                ptrs[i] = MaybeUninit::new(self.get_unchecked_mut(entities[i]));
+            }
+            core::mem::transmute_copy::<_, [&mut V; N]>(&ptrs)
         }
-        core::mem::transmute_copy::<_, [&mut V; N]>(&ptrs)
     }
 }
